feat: read std from toml file, global use of parse args
