+++ Cargo.toml
@@ -61,7 +61,7 @@ clap = { version = "4.5.4", features = ["derive"] }
 eyre = "0.6.12"
 rand = "0.8.5"
 sysinfo = "0.30.8"
-syn = { version = "2.0.66", features = ["full"] }
+syn = { version = "1.0.0", features = ["full"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.68"
 rmp-serde = "1.3.0"
@@ -69,6 +69,7 @@ toml_edit = "0.22.14"
 
 jolt-sdk = { path = "./jolt-sdk" }
 jolt-core = { path = "./jolt-core" }
+common = { path = "./common" }
 
 [profile.test]
 opt-level = 3
