+++ common/Cargo.toml
@@ -8,3 +8,4 @@ ark-serialize = { version = "0.4.2", features = ["derive"] }
 serde = { version = "1.0.193", features = ["derive"] }
 serde_json = "1.0.108"
 strum_macros = "0.25.3"
+syn = { version = "1.0", features = ["full"] }
